rpmsg: glink: Remove IRQF_NO_SUSPEND
Currently interrupts with hardware wake capabilities still cause the
CPUs to handle the interrupt at the GIC layer when IRQF_NO_SUSPEND is
used. This is reproduced in both s2idle and mem suspend.

Remove IRQF_NO_SUSPEND and add enable_irq_wake to properly account for
the wakeup interrupt.